refactor: improve collections handlers code

+++ internal/handlers/collections/delete_collection.go
@@ -2,22 +2,15 @@ package collections
 
 import (
 	"github.com/k4sper1love/watchlist-bot/internal/builders/keyboards"
+	"github.com/k4sper1love/watchlist-bot/internal/builders/messages"
 	"github.com/k4sper1love/watchlist-bot/internal/handlers/states"
 	"github.com/k4sper1love/watchlist-bot/internal/models"
 	"github.com/k4sper1love/watchlist-bot/internal/services/watchlist"
 	"github.com/k4sper1love/watchlist-bot/internal/utils"
-	"github.com/k4sper1love/watchlist-bot/pkg/translator"
 )
 
 func HandleDeleteCollectionCommand(app models.App, session *models.Session) {
-	msg := "⚠️" + translator.Translate(session.Lang, "deleteCollectionConfirm", map[string]interface{}{
-		"Collection": session.CollectionDetailState.Collection.Name,
-	}, nil)
-
-	keyboard := keyboards.NewKeyboard().AddSurvey().Build(session.Lang)
-
-	app.SendMessage(msg, keyboard)
-
+	app.SendMessage(messages.BuildDeleteCollectionMessage(session), keyboards.BuildKeyboardWithSurvey(session))
 	session.SetState(states.ProcessDeleteCollectionAwaitingConfirm)
 }
 
@@ -25,33 +18,22 @@ func HandleDeleteCollectionProcess(app models.App, session *models.Session) {
 	switch session.State {
 	case states.ProcessDeleteCollectionAwaitingConfirm:
 		parseDeleteCollectionConfirm(app, session)
+		session.ClearState()
 	}
 }
 
 func parseDeleteCollectionConfirm(app models.App, session *models.Session) {
-	session.ClearState()
-
-	switch utils.IsAgree(app.Update) {
-	case true:
-		if err := watchlist.DeleteCollection(app, session); err != nil {
-			msg := "🚨 " + translator.Translate(session.Lang, "deleteCollectionFailure", map[string]interface{}{
-				"Collection": session.CollectionDetailState.Collection.Name,
-			}, nil)
-			keyboard := keyboards.NewKeyboard().AddBack(states.CallbackCollectionsManage).Build(session.Lang)
-			app.SendMessage(msg, keyboard)
-			break
-		}
-
-		msg := "🗑️ " + translator.Translate(session.Lang, "deleteCollectionSuccess", map[string]interface{}{
-			"Collection": session.CollectionDetailState.Collection.Name,
-		}, nil)
-
-		app.SendMessage(msg, nil)
-		HandleCollectionsCommand(app, session)
-
-	case false:
-		msg := "🚫 " + translator.Translate(session.Lang, "cancelAction", nil, nil)
-		app.SendMessage(msg, nil)
+	if !utils.IsAgree(app.Update) {
+		app.SendMessage(messages.BuildCancelActionMessage(session), nil)
 		HandleManageCollectionCommand(app, session)
+		return
 	}
+
+	if err := watchlist.DeleteCollection(app, session); err != nil {
+		app.SendMessage(messages.BuildDeleteCollectionFailureMessage(session), keyboards.BuildKeyboardWithBack(session, states.CallbackCollectionsManage))
+		return
+	}
+
+	app.SendMessage(messages.BuildDeleteCollectionSuccessMessage(session), nil)
+	HandleManageCollectionCommand(app, session)
 }
